requirements.txt
torch
torchvision
fastapi
uvicorn
python-multipart
opencv-python
pillow
numpy
